backend/requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.1
django-dotenv==1.4.2
django-filter==25.2
djangorestframework==3.16.0
factory_boy==3.3.3
Faker==37.4.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg==3.2.12
psycopg-binary==3.2.12
requests==2.32.5
ruff==0.14.10
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
